refactor: split ols api and add use client directive
