+++ src/components/widgets/MetadataWidget/BreadcrumbWidget/BreadcrumbWidget.tsx
@@ -5,8 +5,7 @@ import {
   EuiBadge,
   EuiIcon,
 } from "@elastic/eui";
-import { OlsApi } from "../../../../api/OlsApi";
-import { getErrorMessageToDisplay } from "../../../../app/util";
+import { OlsEntityApi } from "../../../../api/ols/OlsEntityApi";
 import { BreadcrumbWidgetProps } from "../../../../app/types";
 import { isEntity } from "../../../../model/ModelTypeCheck";
 import { BreadcrumbPresentation } from "./BreadcrumbPresentation/BreadcrumbPresentation";
@@ -26,7 +25,7 @@ function BreadcrumbWidget(props: BreadcrumbWidgetProps) {
     onNavigateToOntology,
     className,
   } = props;
-  const olsApi = new OlsApi(api);
+  const olsApi = new OlsEntityApi(api);
 
   const { data, isLoading, isSuccess, isError, error } = useQuery(
     ["breadcrumb", api, parameter, entityType, iri, ontologyId, useLegacy],
