+++ src/components/widgets/MetadataWidget/EntityDefinedByWidget/EntityDefinedByWidget.tsx
@@ -1,6 +1,7 @@
+"use client";
+
 import React from "react";
 import { EuiLoadingSpinner, EuiProvider, EuiText } from "@elastic/eui";
-import { OlsApi } from "../../../../api/OlsApi";
 import { QueryClient, QueryClientProvider, useQuery } from "react-query";
 import {
   getErrorMessageToDisplay,
@@ -11,11 +12,12 @@ import { EntityDefinedByPresentation } from "./EntityDefinedByPresentation";
 import ReactDOM from "react-dom";
 import "../../../../style/tssStyles.css";
 import { EntityTypeName } from "../../../../model/ModelTypeCheck";
+import {OlsEntityApi} from "../../../../api/ols/OlsEntityApi";
 
 function EntityDefinedByWidget(props: EntityDefinedByWidgetProps) {
   const { iri, api, parameter, entityType, ontologyId, useLegacy, className } =
     props;
-  const olsApi = new OlsApi(api);
+  const olsApi = new OlsEntityApi(api);
 
   const { data, isLoading, isSuccess, isError, error } = useQuery<{
     ontoList: string[];
