+++ src/components/widgets/MetadataWidget/TabWidget/AlternativeNameTabWidget/AlternativeNameTabWidget.tsx
@@ -1,6 +1,7 @@
+"use client";
+
 import React from "react";
 import { EuiLoadingSpinner, EuiProvider, EuiText } from "@elastic/eui";
-import { OlsApi } from "../../../../../api/OlsApi";
 import { QueryClient, QueryClientProvider, useQuery } from "react-query";
 import { getErrorMessageToDisplay } from "../../../../../app/util";
 import { AlternativeNameTabWidgetProps } from "../../../../../app/types";
@@ -8,11 +9,12 @@ import { Thing } from "../../../../../model/interfaces";
 import { isEntity } from "../../../../../model/ModelTypeCheck";
 import { AlternativeNameTabPresentation } from "./AlternativeNameTabPresentation";
 import ReactDOM from "react-dom";
+import {OlsEntityApi} from "../../../../../api/ols/OlsEntityApi";
 
 function AlternativeNameTabWidget(props: AlternativeNameTabWidgetProps) {
   const { iri, api, parameter, entityType, ontologyId, useLegacy, className } =
     props;
-  const olsApi = new OlsApi(api);
+  const olsApi = new OlsEntityApi(api);
 
   const { data, isLoading, error } = useQuery<Thing>(
     [
