+++ src/components/widgets/MetadataWidget/TabWidget/HierarchyWidget/HierarchyWidget.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { useCallback, useMemo, useReducer } from "react";
 import {
     EuiLoadingSpinner,
@@ -6,7 +8,6 @@ import {
     EuiProvider,
     EuiCard, EuiPanel,
 } from "@elastic/eui";
-import { OlsApi } from "../../../../../api/OlsApi";
 import { Hierarchy, TreeNode } from "../../../../../model/interfaces/Hierarchy";
 import { QueryClient, QueryClientProvider, useQuery } from "react-query";
 import ReactDOM from "react-dom";
@@ -18,14 +19,8 @@ import { randomString } from "../../../../../app/util";
 import { HierarchyWidgetProps, EntityData } from "../../../../../app/types";
 import { isIndividualTypeName } from "../../../../../model/ModelTypeCheck";
 import "../../../../../style/ts4nfdiStyles/ts4nfdiHierarchyStyle.css";
+import {HIERARCHY_WIDGET_DEFAULT_VALUES, OlsHierarchyApi} from "../../../../../api/ols/OlsHierarchyApi";
 
-export const HIERARCHY_WIDGET_DEFAULT_VALUES = {
-  INCLUDE_OBSOLETE_ENTITIES: false,
-  PREFERRED_ROOTS: false,
-  KEEP_EXPANSION_STATES: false,
-  SHOW_SIBLINGS_ON_INIT: false,
-  USE_LEGACY: false,
-} as const;
 
 // TODO: use of entityType has to be reviewed. Currently it is assumed that the entityType of the hierarchy and the specific entity inside it always match (not necessarily true for individual hierarchies, but these have to be reviewed anyways)
 function TreeLink(props: {
@@ -145,13 +140,13 @@ function HierarchyWidget(props: HierarchyWidgetProps) {
   const api: HierarchyBuilder = useMemo(() => {
     switch (backendType) {
       case "ols":
-        return new OlsApi(apiUrl);
+        return new OlsHierarchyApi(apiUrl);
       case "skosmos":
         return new SkosApi(apiUrl);
       case "ontoportal":
         return new OntoPortalApi(apiUrl, apiKey || "");
       default:
-        return new OlsApi(apiUrl);
+        return new OlsHierarchyApi(apiUrl);
     }
   }, [apiUrl, backendType, apiKey]);
 
@@ -319,7 +314,6 @@ function HierarchyWidget(props: HierarchyWidgetProps) {
   return (
     <div className={finalClassName}>
       <EuiPanel
-        layout={"horizontal"}
         style={{ overflowX: "auto", overflowY: "hidden" }}
       >
         {isSuccessHierarchy && hierarchy != undefined ? (
