+++ src/components/widgets/ResourcesWidget/ResourcesWidget.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { ReactNode, useState } from "react";
 import {
   Comparators,
@@ -14,7 +16,6 @@ import {
   EuiCallOut,
 } from "@elastic/eui";
 import { QueryClient, QueryClientProvider, useQuery } from "react-query";
-import { OlsApi } from "../../../api/OlsApi";
 import { css, SerializedStyles } from "@emotion/react";
 import { EuiBasicTableColumn } from "@elastic/eui/src/components/basic_table/basic_table";
 import { OlsResource, ResourcesWidgetProps } from "../../../app/types";
@@ -23,6 +24,7 @@ import ReactDOM from "react-dom";
 import { OLS4Ontology } from "../../../model/ols4-model";
 import { OBO_FOUNDRY_REPO_URL_RAW } from "../../../app/util";
 import "../../../style/ts4nfdiStyles/ts4nfdiResourcesStyle.css";
+import {OlsOntologyApi} from "../../../api/ols/OlsOntologyApi";
 
 const DEFAULT_INITIAL_ENTRIES_PER_PAGE = 10;
 const DEFAULT_PAGE_SIZE_OPTIONS = [10, 25, 50, 100];
@@ -44,7 +46,7 @@ function ResourcesWidget(props: ResourcesWidgetProps) {
     className,
     ...rest
   } = props;
-  const olsApi = new OlsApi(api);
+  const olsApi = new OlsOntologyApi(api);
 
   const [pageIndex, setPageIndex] = useState(0);
   const [pageSize, setPageSize] = useState(initialEntriesPerPage);
